독립적인 기능을 하는 메소드들을 Util 클래스에 담아 따로 분리하는 리팩토링을 함
